Fixed issues of destination folder creation and copy
Easier configuration of destination folder structure and files need to be
copied to the destination.

Now error during creation of the destination structure and during copy of
source files to the destination will be correctly recognized, an error
message will be displayed and the wga pipeline will die.

Fixes issue #4
Fixes issue #5